Feature Addition: Add  Extra Register and Moments
Added a button on the circuit which allows us to increase the number of registers and moments. It makes the circuit more scalable more functional for complex applications. This commit also includes minor content alteration and CSS fix to the Result probability's pre tag and overall width of the website to accomodate higher moment values.